Say the format OpenBSD reads, and skip the limit it has no name for
OpenBSD is the only target here whose int64_t is long long, so PRId64 comes
out "lld" and clang sees four format strings that have never matched their
arguments.  Nothing hides them elsewhere: where int64_t is long, "ld" and
long agree and the check passes.

  jvmciCompilerToVM  displacement is a long printed with JLONG_FORMAT
  jvmciRuntime x2    a jlong printed with INTPTR_FORMAT, which is for
                     pointers; the index is a number, so say JLONG_FORMAT
  jvmciRuntime       a jlong printed with %ld

28 does not carry the first of these -- the message is gone from that tree --
and the port does not touch jvmci at all, so these are 25's own.

has_allocatable_memory_limit asks getrlimit for RLIMIT_AS, which OpenBSD does
not define: it limits the data segment instead and has no address-space
ceiling to report.  Answer that there is no limit, as the print_rlimit above
already does under the same guard.

(cherry picked from commit 8159b5535cd88f05115fd96965e50737b3ee8340)
(cherry picked from commit 82be5af7ec3abdb0b88e863c3549684d12dcf6ab)

+++ src/hotspot/share/jvmci/jvmciCompilerToVM.cpp
@@ -1959,7 +1959,7 @@ C2V_VMENTRY_NULL(jobject, readFieldValue, (JNIEnv* env, jobject, jobject object,
           // enough answer.  It possible to crash in the is_oop call but that just means the crash happens
           // closer to where things went wrong.
           JVMCI_THROW_MSG_NULL(InternalError, err_msg("Read bad oop " INTPTR_FORMAT " at offset " JLONG_FORMAT " in object " INTPTR_FORMAT " of type %s",
-                                                      p2i(value), displacement, p2i(obj()), obj->klass()->external_name()));
+                                                      p2i(value), (jlong) displacement, p2i(obj()), obj->klass()->external_name()));
         }
 
         JVMCIObject result = JVMCIENV->get_object_constant(value);
